requirements.txt
openai>=1.0.0
python-dotenv
requests
colorama
duckduckgo-search
apscheduler
beautifulsoup4
rich
